Add orphan state witnesses to the orphan pool
When the previous block isn't available, we can't process the ChunkStateWitness
and it becomes an orphaned state witness, waiting for the required block to arrive.
In such situations the witness is partialy validated and then put in the `OrphanStateWitnessPool`.
It's impossible to fully validate the witness, as the previous block isn't available,
but we can do some basic checks to curb abusive behavior - check the shard_id, signature,
size, etc.