Add a metric showing how much space nodes and values take up in the recorded Trie (#11761)
Follow up on https://github.com/near/nearcore/pull/11567

Trie is made up of nodes and values referenced by the leaf nodes. Let's
add a metric which shows how much space is taken up by the nodes and the
values in the recorded trie. It'll help us estimate how much overhead
there is from using Merkle trees.

To get the total size I summed the sizes of subtrees for all trie
columns. This is a little bit inaccurate, because it doesn't include the
root node, but it should be good enough, the root node is small so it
doesn't change the results much.

Here's how the trie metrics look like on the current mainnet traffic:

![image](https://github.com/near/nearcore/assets/149345204/b0616400-42a0-4888-94ff-771d8642b872)

![image](https://github.com/near/nearcore/assets/149345204/2ae3eefd-6901-48f4-b876-ea3eb621c529)

You can see a preview here:

https://grafana.nearone.org/d/ddrc1ewgllb0gf/witness-size?var-query0=&var-chain_id=mainnet&var-query0-2=&var-role=dev&var-query0-3=&var-node_type=$__all&var-query0-4=&var-node_id=jan-mainnet-node&var-query0-5=&editIndex=4&var-shard_id=$__all&refresh=1m&from=now-3h&to=now&timezone=browser&showCategory=Bar%20chart