Fix: pop the first bandwidth increase, not the last one
`process_bandwidth_requests` was supposed to try to grant
the first increase from the request, but by mistake I used
Vec::pop(), which pops the last element, not the first one.
Fix it by using `VecDeque` and `pop_front`.

I verified that it doesn't make the performance worse. VecDeque is a
Vec-based ring buffer, so it's very similar to popping items from a
normal Vec.