Block postprocessed check
Usually the prepare transactions job runs in parallel with chunk application,
before the block that contains the applied chunk is postprocessed.

However in rare cases (weird thread scheduling, testloop reordering things)
it might happen that the job starts after the block is postprocessed.

In such cases it's better to discard the job and prepare transactions the
normal way, in `produce_chunk` which happens right after postprocessing the block.

This is because the job is not aware of the block that was just postprocessed.
If we run the job, there's a risk that transactions that were created using
the latest postprocessed block will be rejected because the job isn't
aware of the latest block. This happens is some testloop tests and makes
them fail.

Let's disable the job when this edge case happens.