Use OptimisticBlockKeySource in CachedShardUpdateKey constructor
CachedShardUpdateKey is currently created based on
`ApplyChunkBlockContext`, but the constructor only reads
a few fields from the block context.

In other situations (e.g. early transaction preparation)
we might not have the whole `ApplyChunkBlockContext` and we
would like to create the `CachedShardUpdateKey` using only
the data that it really needs, without creating a whole
`ApplyChunkBlockContext`.

Let's directly use OptimisticBlockKeySource in the constructor.
It contains the block data needed to create a `CachedShardUpdateKey`
without any other extra fields.